requirements.txt
absl-py==2.1.0
asttokens==2.4.1
blinker==1.8.2
blobfile==2.1.1
blurhash==1.1.4
boost==0.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.1
comm==0.2.2
contourpy==1.2.0
cvzone==1.6.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dlib==19.24.99
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.13.1
filterpy==1.4.5
Flask==3.0.3
fonttools==4.49.0
fsspec==2024.2.0
greenlet==3.0.3
grpcio==1.62.1
httpie==3.2.3
huggingface-hub==0.24.1
idna==3.7
imageio==2.34.0
ipympl==0.9.3
ipython==8.22.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
lxml==4.9.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
Mastodon.py==1.8.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycryptodomex==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.7.21
timm==1.0.7
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.2.62
ultralytics-thop==2.0.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
